requirements.txt
google-genai>=0.3.0
pydantic>=2.5.0
pyyaml>=6.0
numpy>=1.26
python-dotenv>=1.0.0
playwright>=1.40.0
Pillow>=10.0.0
imageio>=2.34.0
